tenants/e-commerce/pyproject.toml
[project]
name = "e-commerce"
version = "0.1.0"
description = "This is an application for Poddle debug testing the observability in case users instrument their code"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "pydantic>=2.12.5",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
]
